requirements.txt
requests~=2.32.3
pandas~=2.2.3
protobuf~=5.29.1
python-dotenv~=1.0.1